fix: Add warning when running `bench setup requirements` in wrong directory
The bench setup requirements command must be run in the root directory of the
bench to work correctly. This commit adds a warning message when the number of
detected apps is equal to zero, which means that the command might have been
run in a descendant directory of the bench root.